Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

This driver works very well on my galaxy5 unofficial port using a AR6003 chip. I also suggest that you submit this for the gingerbread branch (also tested and working perfectly with no changes).

Jerpelea, I'm not sure if your -1 is very helpful. The v6 supplicant is still provided in the source, so it should be maintained. Someone can forward-port this driver to v8 later if they want.

Patch-set: 2